requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
imageio==2.36.1
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-image==0.25.0
scipy==1.14.1
setuptools==75.1.0
six==1.17.0
SQLAlchemy==2.0.36
tifffile==2024.12.12
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
wheel==0.44.0
xgboost==2.1.3
